+++ src/main/java/roomescape/reservation/Reservation.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.reservation;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
