+++ src/main/java/roomescape/reservation/dao/JdbcReservationDao.java
@@ -1,10 +1,11 @@
-package roomescape;
+package roomescape.reservation.dao;
 
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Component;
+import roomescape.reservation.Reservation;
 
 @Component
 public class JdbcReservationDao implements ReservationDao {
