+++ src/main/java/roomescape/reservation/dao/ReservationDao.java
@@ -1,6 +1,7 @@
-package roomescape;
+package roomescape.reservation.dao;
 
 import java.util.List;
+import roomescape.reservation.Reservation;
 
 public interface ReservationDao {
     List<Reservation> findAll();
