+++ src/main/java/roomescape/reservation/web/ReservationRequest.java
@@ -1,7 +1,8 @@
-package roomescape;
+package roomescape.reservation.web;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
+import roomescape.reservation.Reservation;
 
 public record ReservationRequest(String name, LocalDate date, LocalTime time) {
     public Reservation toReservation() {
